refactor: split map protocol module
